# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ecphoria Context Lake API
  description: >
    REST API for Ecphoria — episodic events, semantic search, agent state, and the bi-temporal
    memory cognition layer. When `auth_enabled`, all `/api/v1/*` routes require
    `Authorization: Bearer <token>` (API key, JWT HS256, or OIDC RS256) and are tenant-scoped.
  version: "0.2.0"
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8432
    description: Local
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/memories:
    post:
      summary: Add a memory (dedup / contradiction / importance)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
                subject: { type: string }
                importance: { type: number }
                user_id: { type: string }
                agent_id: { type: string }
                session_id: { type: string }
                metadata: { type: object }
      responses: { '200': { description: MemoryAdd (memory + outcome) } }
    get:
      summary: List active memories in a scope
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: user_id, in: query, schema: { type: string } }
        - { name: agent_id, in: query, schema: { type: string } }
        - { name: session_id, in: query, schema: { type: string } }
      responses: { '200': { description: '{ memories: [...], count: N }' } }
...
  /api/v1/admin/tenants/{tenant_id}:
    delete:
      summary: GDPR erasure — delete all data for a tenant across every store (admin)
      parameters: [ { name: tenant_id, in: path, required: true, schema: { type: string } } ]
      responses: { '200': { description: Per-store deletion summary } }
...